Moved add compute_update_indices_sparse and compute_update_indices_dense to inplace_utils.py
